pyproject.toml
[project]
name = "transactions-analyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "Galina Mordovina",email = "glukoloid@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "finnhub-python (>=2.4.24,<3.0.0)"
]


[build-system]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
flake8 = "^7.3.0"
black = "^25.1.0"
